project/requirements.txt
torch
torchvision
numpy
matplotlib
pillow
sympy
fastapi
uvicorn
